chore: move tm functions to stdlib package (#758)
